fix: refactor rowLinkWrapper
